notifications -> mypage 패키지변경
