+++ app/src/main/java/com/example/zerobin/ui/mypage/MyPageStampFragment.kt
@@ -1,6 +1,5 @@
-package com.example.zerobin.ui.notifications
+package com.example.zerobin.ui.mypage
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -9,18 +8,18 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.example.zerobin.R
-import com.example.zerobin.databinding.FragmentMyPageShopBinding
-import com.example.zerobin.ui.home.adapter.ShopAdapter
-import com.example.zerobin.ui.home.shop.ShopDetailActivity
+import com.example.zerobin.databinding.FragmentMyPageStampBinding
+import com.example.zerobin.model.Review
+import com.example.zerobin.ui.review.adapter.ReviewAdapter
 
 
-class MyPageShopFragment : Fragment() {
+class MyPageStampFragment : Fragment() {
 
-    private lateinit var binding: FragmentMyPageShopBinding
 
+    private lateinit var binding: FragmentMyPageStampBinding
     private lateinit var myPageViewModel: MyPageViewModel
 
-    private val shopAdapter by lazy { ShopAdapter() }
+    private val myPageStampAdapter by lazy { ReviewAdapter() }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -30,9 +29,9 @@ class MyPageShopFragment : Fragment() {
         myPageViewModel = ViewModelProvider(this).get(MyPageViewModel::class.java)
 
         binding =
-            DataBindingUtil.inflate(inflater, R.layout.fragment_my_page_shop, container, false)
+            DataBindingUtil.inflate(inflater, R.layout.fragment_my_page_stamp, container, false)
         binding.lifecycleOwner = viewLifecycleOwner
-        binding.vm = myPageViewModel
+        binding.reviewVM = myPageViewModel
 
         return binding.root
     }
@@ -43,16 +42,23 @@ class MyPageShopFragment : Fragment() {
         setShopAdapter()
         observeLiveData()
         requestShopList()
-        setListener()
+        //setListener()
     }
 
     private fun setShopAdapter() {
-        binding.shopRecyclerView.adapter = shopAdapter
+        binding.myPageStampRecyclerView.adapter = myPageStampAdapter
     }
 
     private fun observeLiveData() {
         myPageViewModel.myUser.observe(viewLifecycleOwner) {
-//            shopAdapter.setItem(it.favoriteShop)
+            val item = ArrayList<Review>()
+            for (i in 0 until it.ReviewList.size) {
+                if (it.ReviewList[i].favorite) {
+
+                    item.add(it.ReviewList[i])
+                }
+            }
+            myPageStampAdapter.setItem(item)
 
         }
     }
@@ -61,13 +67,13 @@ class MyPageShopFragment : Fragment() {
         myPageViewModel.requestMyPage()
     }
 
-    private fun setListener() {
-        shopAdapter.onClick = { shop ->
+/*    private fun setListener() {
+        myPageStampAdapter.onClick = { shop ->
             val intent = Intent(requireContext(), ShopDetailActivity::class.java).apply {
                 putExtra(ShopDetailActivity.EXTRA_SHOP, shop)
             }
             startActivity(intent)
         }
-    }
+    }*/
 
 }
