+++ app/src/main/res/layout/fragment_my_page.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="userVM"
-            type="com.example.zerobin.ui.notifications.MyPageViewModel" />
+            type="com.example.zerobin.ui.mypage.MyPageViewModel" />
 
 
     </data>
@@ -15,7 +15,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.notifications.MyPageFragment">
+        tools:context=".ui.mypage.MyPageFragment">
 
         <TextView
             android:id="@+id/myPageDetail"
