+++ app/src/main/res/layout/fragment_my_page_review.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="review_VM"
-            type="com.example.zerobin.ui.notifications.MyPageViewModel" />
+            type="com.example.zerobin.ui.mypage.MyPageViewModel" />
 
     </data>
