+++ app/src/main/res/layout/fragment_my_page_shop.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="VM"
-            type="com.example.zerobin.ui.notifications.MyPageViewModel" />
+            type="com.example.zerobin.ui.mypage.MyPageViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
