+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -29,7 +29,7 @@
 
     <fragment
         android:id="@+id/navigation_my_page"
-        android:name="com.example.zerobin.ui.notifications.MyPageFragment"
+        android:name="com.example.zerobin.ui.mypage.MyPageFragment"
         android:label="@string/title_my_page"
         tools:layout="@layout/fragment_my_page" />
 </navigation>
